♻️ Refactor: entity 수정 validation 추가, dto id 삭제
