--- src/main/java/roomit/web1_2_bumblebee_be/domain/workplace/test.java
@@ -1,4 +1,0 @@
-package roomit.web1_2_bumblebee_be.domain.workplace;
-
-public class test {
-}
